refactor - cleanup UI types in but-workspace a bit
