+++ crates/but-workspace/src/changeset.rs
@@ -6,11 +6,10 @@
 
 use crate::{
     RefInfo,
-    ref_info::{
-        ui,
-        ui::{LocalCommit, LocalCommitRelation},
+    ui::{
+        PushStatus,
+        ref_info::{LocalCommit, LocalCommitRelation},
     },
-    ui::PushStatus,
 };
 use bstr::{BStr, BString, ByteSlice, ByteVec};
 use but_core::{ChangeState, commit::TreeKind};
@@ -94,7 +93,7 @@ impl RefInfo {
             repo,
             upstream_commits.iter().filter_map(|id| {
                 but_core::Commit::from_id(id.attach(repo))
-                    .map(ui::Commit::from)
+                    .map(crate::ui::ref_info::Commit::from)
                     .ok()
             }),
             cost_info,
@@ -262,7 +261,7 @@ impl PushStatus {
 
 fn changeset_identifier(
     repo: &gix::Repository,
-    commit: Option<&ui::Commit>,
+    commit: Option<&crate::ui::ref_info::Commit>,
     elapsed: &mut Duration,
 ) -> anyhow::Result<Option<Identifier>> {
     let Some(commit) = commit else {
@@ -296,7 +295,7 @@ enum ChangeId {
 
 fn lookup_similar<'a>(
     map: &'a Identity,
-    commit: &ui::Commit,
+    commit: &crate::ui::ref_info::Commit,
     expensive: Option<&Identifier>,
     change_id: ChangeId,
 ) -> Option<&'a gix::ObjectId> {
@@ -312,7 +311,7 @@ fn lookup_similar<'a>(
 /// Returns the fully-loaded commits suitable to be passed to UI, to have better re-use.
 fn create_similarity_lut(
     repo: &gix::Repository,
-    commits: impl Iterator<Item = impl Borrow<ui::Commit>>,
+    commits: impl Iterator<Item = impl Borrow<crate::ui::ref_info::Commit>>,
     (max_commits, num_tracked_files): (usize, usize),
     expensive: bool,
 ) -> anyhow::Result<Identity> {
@@ -633,7 +632,7 @@ enum Identifier {
     ChangesetId(ChangesetID),
 }
 
-fn commit_data_id(c: &ui::Commit) -> anyhow::Result<Identifier> {
+fn commit_data_id(c: &crate::ui::ref_info::Commit) -> anyhow::Result<Identifier> {
     let mut hasher = gix::hash::hasher(gix::hash::Kind::Sha1);
 
     let gix::actor::Signature {
